1. If the execution of an SQL query throws an Exception the logged warning message will now include the concrete query parameters, which can be very useful for future application debugging. 2. Added the pool name, pool taken and pool remainingCreated parameters when logging important messages at warn or error level in the  format "... from pool NAME (taken/remainingCreated) ...". The last could be quite useful for debugging purposes.

+++ src/main/java/org/vibur/dbcp/pool/PoolOperations.java
@@ -84,6 +84,8 @@ public class  PoolOperations {
     }
 
     private class PoolReducer extends SamplingPoolReducer {
+        private final Logger logger = LoggerFactory.getLogger(PoolReducer.class);
+
         private PoolReducer(PoolService poolService, long timeInterval, TimeUnit unit, int samples) {
             super(poolService, timeInterval, unit, samples);
         }
@@ -95,8 +97,8 @@ public class  PoolOperations {
                 if (!(thrown instanceof ViburDBCPException))
                     terminate();
             } else
-                logger.debug("Pool {}, intended reduction {} actual {}, currently taken {} and remainingCreated {}",
-                        config.getName(), reduction, reduced, pool.taken(), pool.remainingCreated());
+                logger.debug("Pool {} ({}/{}), intended reduction {} actual {}.",
+                        config.getName(), pool.taken(), pool.remainingCreated(), reduction, reduced);
         }
 
         public String toString() {
@@ -161,7 +163,8 @@ public class  PoolOperations {
     }
 
     public void terminate() {
-        poolReducer.terminate();
+        if (poolReducer != null)
+            poolReducer.terminate();
         pool.terminate();
     }
